refactor(cli): moving binary target related files onto src/bin dir
