+++ sn_cli/src/bin/subcommands/files.rs
