+++ sn_cli/src/bin/subcommands/register.rs
