+++ sn_cli/src/bin/subcommands/wallet.rs
@@ -5,6 +5,7 @@
 // under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
+
 mod audit;
 pub(crate) mod helpers;
 pub(crate) mod hot_wallet;
