+++ sn_cli/src/bin/subcommands/wallet/audit.rs
