+++ sn_cli/src/bin/subcommands/wallet/helpers.rs
