+++ sn_cli/src/bin/subcommands/wallet/hot_wallet.rs
