+++ sn_cli/src/bin/subcommands/wallet/wo_wallet.rs
