+++ sn_cli/src/lib.rs
